Code cleanup and refactor of LCD tests
